Assign same ID to leaves with the same label
The code does not check that two leaves in the same tree don't have the
same label.  This would likely wreak havoc on the current implementation
and is something we shouldn't support because it would add complications
and should simply not occur with the inputs we care about.  For leaves
with the same labels in different trees, it is imperative that they have
the same IDs because we use IDs to look up leaves in different trees
that correspond to each other.