Clean up tree implementation
Instead of pointing to the first child and linking the children
in a doubly-linked list, we now have every node store a vector of
its children.  This gets rid of a lot of pointers that need to be
maintained when manipulating trees.  For high-degree nodes, this
is less efficient because specifically child deletions no longer
take constant time.  As long as there are few children, which is
currently the case because we only support binary trees, the constant
factor savings are likely to outweigh this.