Reduce repeated exploration of redundant branches
If one branch resolves a non-trivial cherry (a,b) and then the cherry
(c,d), then it is useless to resolve try resolving the cherry (a,b)
again in the branch that starts by resolving (c,d) unless (a,b) now
occurs in more trees than before resolving (c,d).  The new
implementation keeps track of the number of trees containing a cherry
the last time this cherry was cut, so it shouldn’t affect the
correctness but should reduce the exploration of redundant branches
significantly.