Add command line flags to toggle optimizations
Two optimizations can now be turned on/off using command line flags:

-b: Disables bounding the branching fan-out based on the current
    hybridization number

-r: Enables the elimination of redundant branches based on edge
    cuts in previously explored branches.

We proved that bounding the fan-out is correct, so this is on by
default but can be disabled to double-check that it works and to
evaluate its effectiveness.

We haven’t proven yet that the redundant branch optimization is
correct, so it’s off by default but can be enabled to test whether
it seems to work and to evaluate its effectiveness.