Implement work stealing
This is not a real work stealer but rather "work donator".  Worker
threads that are not idle check for idle threads, then hive off
some of their work and send it to an idle thread.  This seemed
easier to implement than a proper work stealer.  rayon supports
work stealing but requires tasks to be represented as sendable
data packages, something that doesn't seem to be easy to integrate
with the "do not copy, modify and undo" approach used by the search.