Fix redundant branch optimization
The old optimization was equivalent to allowing a cherry (v,w) to be moved
before a cherry (u,v) in the tree-child sequence, which results in a sequence
that is no longer tree-child.  The new optimization forbids this by resetting
the cut count of v to zero in every non-trivial cherry when we add the pair
(u,v) to the sequence.  This has the effect that the any subsequent branch that
inspects (v,w) does not skip this branch, that is, the optimization no longer
assumes that (v,w) could have been moved earlier in the sequence.