Add logic to count how many times a leaf has been pruned
We count how many times a leaf has been pruned and refuse to add a
non-trivial cherry if the leaf has already been pruned, if we are
looking for a binary tree-child sequence.

Still need to add some logic at the top level that now allows the whole
search to fail, no matter what the parameter is, which can happen for
binary networks.