Implement new default behaviours
We now have a proof that redundant branch elimination is correct, so it's no
longer experimental and should be on by default.  Thus, -r now turns it off
instead of on.

Our experiments suggest that 2 threads per core and -w 100 are sane defaults.
So that's what the absence of these options translates to now.

The whole point is that -p, -w, -r, -c should only be used by users who want to
experiment with the algorithm itself.  For anybody else, not providing any
options should offer a configuration that is likely to perform well.